+++ pkg/geo/data/embed.go
@@ -1,0 +1,8 @@
+package data
+
+import (
+	_ "embed"
+)
+
+//go:embed geodata.bin
+var GeoData []byte